# pkg/estimator/api/estimator-types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package api

const (
	ApiKeyAuthScopes = "apiKeyAuth.Scopes"
)

// PowerConsumption defines model for PowerConsumption.
type PowerConsumption struct {
	// CpuMilli The amount of CPUs required by each workload.
	CpuMilli int `json:"cpu_milli"`

	// NumWorkload The amount of workloads have to be allocated.
	NumWorkload int `json:"num_workload"`

	// WattIncrease The estimated increase in power cunsumption for each workload.
	WattIncrease *[]int `json:"watt_increase,omitempty"`
}

// PostNamespacesNamespaceEstimatorsEstimatorResourcesPowerconsumptionJSONRequestBody defines body for PostNamespacesNamespaceEstimatorsEstimatorResourcesPowerconsumption for application/json ContentType.
type PostNamespacesNamespaceEstimatorsEstimatorResourcesPowerconsumptionJSONRequestBody = PowerConsumption